Refactor DomainQL meta data handling
So far we've ad hoc added some more fields to our __schema introspection
result and it has gotten a bit out of hand and is only getting bigger.

This formalizes the approach to meta data in GraphQL and modularizes it
so that outside participients can contribute.

At the core we have the DomainQLMeta class which encapsulates all meta
data. It containes a "types" property that contains the map of
DomainQLTypeMeta instances containing the meta data for the GraphQL
types.

A MetadataProvider interface can be implemented to add three kinds of
meta data

 * a user-named addendum to the root level (This refactors "genericTypes"
   and "relations" to such addenda)

 * user-named type level meta data ("nameFields" was refactored to
   use this)

 * user-named field level meta data (to be used for "computed" and field
   lengths and decimal precision etc in automaton)

Adapts the existing meta data tests to the new system.