SchemaDataProvider: Change default name
With the new {schema,meta} structure, it is awkward to have
schema.schema. Schema was mostly an apt name because it *was* the
GraphQL schema which now moved one down.

And while we're just changing the meta data structure anyway, it
shouldn't be any more breaking and in fact clearer that some large scale
changes happened.

This changes the default js view name to "domainQL".