DomainQL: Update internal data on overloaded types
Our way of using JOOQ generated database types and fields to
conveniently define our relations with autocompletion has the drawback
that it stores the original JOOQ types as types in the relation.

The same happens with the TableLookup classes (which might be a
refactoring target to be merged with OutputType)

This updates type references for overloaded types so that there is no
type mismatch anymore.