Meta: Remove name from type meta
Removes the name property from the type meta. Sometimes it can be a good
thing to have key of a map as reduntant property on the elements. Mostly
because you might want to pass around the element without having to take
care of the key separately.

Passing around meta data structures makes no sense though.