Implement ComputedMetaDataProvider
So using GraphQL directives for the "computed" status of a field is a
bust because while we can access that directive on the Java side, we
cannot get it with GraphQL schema introspection.

This implements a ComputedMetadataProvider that checks the Java types of
all GraphQL output types for the @GraphQLComputed annotation and then
sets the field meta "computed" to true for that field.

Adds a test to check that behavior.