RelationModel: Add meta tags
With the existing meta data system, we have a flexible way of adding
domain meta data to our domain. The only thing that's a bit awkward is
the relation models which we added as their own built-in addendum.

Adding the meta data directly where we configure our relation allows for
convenient configuration where we don't have to refer back to relation
ids to configure the meta but can do so right in the builder.

Adds a little test.