Upgrade GraphQL to 20.0
The recent GraphQL version changed the Scalar Type handling.
GraphQLScalarType has a private constructor and needs to be created via
GraphQLScalarType.newScalar().

This throws all our GraphQLScalar implementation into disarray. We solve
this by changing all our GraphQLScalar implementation to be Util classes
with a single builder method that creates the GraphQLScalarType with the
appropriate name, description and coercing.

Changes the coercings to be a static inner class of the Util classes.

The GraphQL project retired some scalar types. We need "Byte" and "Long"
however, so we recreate them.

Harmonizes the scalar names not to have a GraphQL prefix.

The fact that GraphQLScalarType is no longer extendable also means that
DomainQLAware now only works for the Coercing-Implementation. Luckily
that is exactly where the knowledge of the schema/domainQL instance is
needed.