Upgrade to Java 17
Nothing too complicated.

Formally adds the missing javax-annotation as dependency.

Removes the PreloadedGraphQLQueryProvider. Not sure if anyone ever
actually used that one. It was introduced as alternative to the static
analysis based GraphQL injection in automaton. But it is based on
Nashorn which no longer exists, so... If you want to do something
similar I guess you'd use GraalVM for this now.