Update DomainQL logic to jakarta.*
I did the wrong thing initially and added the javax.* annotation that
were missing back to pom. With the old generated domain classes this
worked fine but totally missed the point of modernization.

This commit removes the old JPA and java validation API again and adds
the new jakarta.* ones, because that's what the new Version of JOOQ
generates and what we now have to deal with.

Changes the framework code to only regard jakarta.* annotations.

    jakarta.persistence.Column
    jakarta.persistence.Table
    jakarta.validation.constraints.NotNull

Updates the generated classes used in Tests to the new format.

Fixes confusing documentation for one test.