Adapt code to new GraphQL input behavior
We used to only receive the defined properties of our input objects to
be from the GraphQL environment.

Now we receive null values for undefined properties.

This can be prevented by annotating the property getter with

    @GraphQLField( defaultValue = "value" )

This commit fixes/improves two other problems that surfaced due to the
new behavior.

Enables to correctly deal with null lists.

Provides a better error description when the system is trying to write a
null value into a primitive type property.