Revert "Fixed wrong Timestamp-Conversion (#14)"
This reverts commit 00bef5d935a90e592ec1dc97fce158d1306472b1.

While the change did help fix some bugs downstream, it violates the
central promise of scalar conversion in that the output in one direction
of course must be a valid input in the other

e.g. "15" -> 15 - > "15"

Here the literal Z suffix is exactly what Instant.parse wants and what
is inherently in line with the meaning of the Timestamp type, i.e. a
type that is without timezone.